Move listener to a dedicated module
We're going to introduce a new listener for kinto-changes, and that
will probably be the more important one for our use case, but it's
still worth keeping the old one around in case it's of use to someone.
