+++ tests/conftest.py
@@ -21,7 +21,7 @@ def megaphone_settings():
     return {
         'project_name': 'kinto megaphone test',
         'event_listeners': 'mp',
-        'event_listeners.mp.use': 'kinto_megaphone.listeners',
+        'event_listeners.mp.use': 'kinto_megaphone.listeners.collection_timestamp',
         'event_listeners.mp.api_key': 'token',
         'event_listeners.mp.url': 'http://megaphone.example.com',
         'event_listeners.mp.broadcaster_id': 'bcast',
